# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/species": {
      "get": {
        "operationId": "SpeciesController_listSpecies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or credentials are invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user does not have permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the animal species catalog",
        "tags": [
          "species"
        ]
      }
    },
    "/api/v1/species/{id}/breeds": {
      "get": {
        "operationId": "SpeciesController_listBreeds",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Species id (UUID).",
            "schema": {
              "format": "uuid",
              "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreedListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required or credentials are invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated user does not have permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the breeds for a species",
        "tags": [
          "species"
        ]
      }
    },
...
  },
  "info": {
    "title": "Refugiapp API",
    "description": "API para la gestion de refugios de animales.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "code": {
            "type": "string",
            "example": "RESOURCE_CONFLICT"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "User already exists.",
            "description": "Human readable error message or a list of validation messages."
          },
          "error": {
            "type": "string",
            "example": "Conflict"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-07T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/users"
          },
          "requestId": {
            "type": "string",
            "example": "7c0c5b2a-3f8e-4b1a-9d2e-1a2b3c4d5e6f"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "error",
          "timestamp",
          "path"
        ]
      },
...
      "SpeciesResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "slug": {
            "type": "string",
            "example": "dog"
          },
          "labelEs": {
            "type": "string",
            "example": "Perro"
          }
        },
        "required": [
          "id",
          "slug",
          "labelEs"
        ]
      },
      "SpeciesListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpeciesResponseDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "BreedResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "speciesId": {
            "type": "string",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "slug": {
            "type": "string",
            "example": "labrador-retriever"
          },
          "labelEs": {
            "type": "string",
            "example": "Labrador retriever"
          }
        },
        "required": [
          "id",
          "speciesId",
          "slug",
          "labelEs"
        ]
      },
      "BreedListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BreedResponseDto"
            }
          }
        },
        "required": [
          "items"
        ]
      }
    }
  }
}