# src/modules/species/interfaces/dto/breed-response.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class BreedResponseDto {
  @ApiProperty({ example: 'f47ac10b-58cc-4372-a567-0e02b2c3d479' })
  id!: string;

  @ApiProperty({ example: 'f47ac10b-58cc-4372-a567-0e02b2c3d479' })
  speciesId!: string;

  @ApiProperty({ example: 'labrador-retriever' })
  slug!: string;

  @ApiProperty({ example: 'Labrador retriever' })
  labelEs!: string;
}